feat(proto-build): improve modularity
