+++ proto-build/src/parser/consts.rs
@@ -1,0 +1,2 @@
+// prost Message implements a generic B which was conflicting with this script
+pub const GENERICS: [&str; 10] = ["A", "BB", "C", "D", "E", "F", "G", "H", "I", "J"];
